changed pipeline model to be able to accept multiple imports
