+++ pkg/models/job.go
@@ -13,30 +13,6 @@ type SecretsRef struct {
 	FileReference *FileReference `json:"file_reference,omitempty"`
 }
 
-type SourceType string
-
-const (
-	SourceTypeLocal  SourceType = "local"
-	SourceTypeRemote SourceType = "remote"
-)
-
-type ImportSource struct {
-	SCM          Platform   `json:"scm,omitempty"`
-	Organization *string    `json:"organization,omitempty"`
-	Repository   *string    `json:"repository,omitempty"`
-	Path         *string    `json:"path,omitempty"`
-	Type         SourceType `json:"type,omitempty"`
-}
-
-type Import struct {
-	Source      *ImportSource  `json:"source,omitempty"`
-	Version     *string        `json:"version,omitempty"`
-	VersionType VersionType    `json:"version_type,omitempty"`
-	Pipeline    *Pipeline      `json:"pipeline,omitempty"`
-	Parameters  map[string]any `json:"parameters,omitempty"`
-	Secrets     *SecretsRef    `json:"secrets,omitempty"`
-}
-
 type Job struct {
 	ID                   *string                  `json:"id,omitempty"`
 	Name                 *string                  `json:"name,omitempty"`
